changed int and double to float for x,y and speed

+++ src/source/Caracter.hpp
@@ -16,7 +16,7 @@ protected:
     float y;
     Direction direction = WEST;
     Direction nextDirection;
-    double speed;
+    float speed;
 
     bool canMove(Map* map,int i,int j);
